Socket reading code is in.
The client code can now open, read, and parse data, passing it to callbacks. Still untested. Next up is working on more protocol spec, fixing things I may have missed, and more of the spec that I'm trying to write without knowing a thing of what the rest of the team expects of it.