Added small vector maths library, created positionable.h, added my plans for IL
Vector math library is a modified version of the one from sCenegraph. Positionable.h is a base-class for 3-dimensional objects, which is extended by classes like drawable3d (which have a pointer back to the positionable). Plans for IL are in the form of badly drawn paint diagrams, but the general structure is pretty obvious in them. The structure reduces the amount of interdependence on other header files, which is nice.