Add common/log.h, clean up warnings in graphics
Log function was added to control the verbosity of text being spammed into
stdout (and to change stdout to a text file if needed). Updated build
script to define DEBUG variable for now, in the future there will be an
option for it. Graphics module has had warnings cleaned up and I removed
an unnecessary pointer in the heightmap struct. Event.h now uses types
declared in stdint.h, and has an undefined-length-array at the end of
il_Event_Event to make adding data easier.