Switched to Lua 5.2 and disabled require+debug
You will have to install lua 5.2 and make sure that the headers and
shared objects are available.

If you keep getting linker errors, then putting the 5.2 header files in
the src directory and putting a symlink for liblua in the project root
should fix it. This may or may not require modifying your lua.pc in your
pkgconfig directory.

This change was made because of several features that have been made
available in lua 5.2 that aid implementation, and help make more secure
interfaces.

The debug API and require function have been disabled (when security
contexts are introduced, they will be added back). The reason for this
is there is essentially no amount of securing of the engine interface
that can be done that would prevent a potential exploit while the debug
interface is available. The require function had to be disabled as well,
because you could have gotten the debug interface (or any other similar
library) with minimal effort (`require 'debug'`).