Finished Lua for varying definitions of 'finished'
Things left to do to the lua api to make it nicer: Upping the
intelligence, making things more readable, documenting the whole
interface, use some more of the new 5.2 functions that are nice.

Warning: The API has not been thoroughly tested. If you encounter an
issue, it's most likely my fault.