Arrays&lists, graphics code modified
Arrays and linked lists should have been added a long time ago. They are
used in many places throughout the engine yet I have never decided to
make the boilerplate located in one single place. Now that this is
cleaned up, I can move things to array types and get various speed
boosts where necessary.

The changse to the graphics API are part of an undergoing process of me
deciding what to do with the graphics module to get it functioning as
quickly as possible without various opengl tricks and microopts.