Removed old script API, setup luaJIT FFI stuff
It'll be easier to create and maintain a lua API this way, and make the
lua/C boundary more efficient. Not to mention the huge performance boost
just from using a JIT'd lua.

Unfortunately, this means that there is no more lua 5.2 support until
luaJIT updates and adds API compatibility with 5.2, which as far as I
know will not be happening for some time.