Fixed issue with matrix.lua
As it turns out, you can't define a type twice in your lua application.
Ever. So you have to make a file which has the definition, and require
that from all your files (This is what scalar_defs.lua is for).

Matrix interface is fleshed out, and the test script now sets a
perspective on the camera. Still need to bind some sort of keyboard
handling, not sure how C-based I want that to be.