Events!
There is once again a way to bind event handlers from within lua.
Additionally, the demo has been updated to demonstrate this. There is
now simple camera movements.

Additionally, several memory leaks with the OBJ loader have been fixed,
and IL uses about 6M of memory after it has finished initializing. Most
of that is mesa/xorg/etc., I couldn't find anything that was one of my
allocations larger than massif's threshold (1%).

This is probably the point at which is is possible to sort of kind of
write a game using the engine, that is, if you don't care for sounds,
animations, networking, physics, lighting, among other things. You can
load a mesh file, texture it, position it however you want, apply
arbitrary glsl shaders to it, move the camera around, and so on.