Restructure for easier embedding into a larger project tree.
Should not require any modifications if the tiffgl exists on your path
somewhere already, otherwise you will have to create a Tup config which
contains the path.

If you are embedding tiffgl higher up in a project tree, you will want
to add -L and -I for it to LDFLAGS and CCFLAGS, respectively. IL will
automatically pass -ltgl. You will also have to set $TGL_LIB to point to
the location of libtgl.a, for order-only dependencies.